Changing FFmpeg-service to a separate file
